Re-submission with corrected deliverables
Fixed number of drivers and avg fare per driver variables by using correct dataframe to sum the columns with (city_data_df not the newly merged pyber_data_df). Once I had the correct values I ran the steps again and created the correct dataframe in step 8. Finally, for deliverable 2 I used the [df.plot] matplotlib function as opposed to the [ax, fig = subplots()] matplotlib function I originally used in order to correct the values on the x-axis (months as opposed to the individual week) and resized the figure to more closely match the figure we were provided in the challenge.